move middleware and auth config to @io-ipatente/core
