+++ apps/vehicles/src/lib/bff/with-jwt-auth-voucher-handler.ts
@@ -1,7 +1,7 @@
 import { Voucher } from "@io-ipatente/core";
 import { NextRequest, NextResponse } from "next/server";
+import { User } from "next-auth";
 
-import { CustomUser } from "../../../types/next-auth";
 import { withJWTAuthHandler } from "./with-jwt-auth-handler";
 import { withVoucherHandler } from "./with-voucher-handler";
 
@@ -12,7 +12,7 @@ export const withJWTAuthAndVoucherHandler = (
       additionalDataJWS: string;
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       params: any;
-      user: CustomUser;
+      user: User;
       voucher: Voucher;
     },
   ) => Promise<NextResponse> | Promise<Response>,
