+++ apps/vehicles/src/middleware.ts
@@ -1,30 +1,7 @@
 import { auth } from "@/auth";
-import { NextRequest, NextResponse, userAgent } from "next/server";
+import { handleAuthRequest } from "@io-ipatente/core";
 
-const isBrowserContext = (request: NextRequest) =>
-  !!userAgent(request).browser.name;
-
-export async function middleware(request: NextRequest) {
-  if (request.nextUrl.pathname === "/api/auth/callback/fims") {
-    if (!isBrowserContext(request)) {
-      const redirectUrl = request.nextUrl;
-      redirectUrl.pathname = "/api/auth/callback/fims/cookies";
-      request.cookies
-        .getAll()
-        .forEach((cookie) =>
-          redirectUrl.searchParams.set(cookie.name, cookie.value),
-        );
-      return NextResponse.redirect(redirectUrl);
-    }
-  } else {
-    const session = await auth();
-    if (!session) {
-      return NextResponse.rewrite(
-        new URL("/api/auth/signin", request.nextUrl.origin),
-      );
-    }
-  }
-}
+export default auth(handleAuthRequest);
 
 export const config = {
   matcher: [
