+++ packages/core/package.json
@@ -33,22 +33,34 @@
     "@pagopa/eslint-config": "^4.0.1",
     "@types/jsonwebtoken": "^9.0.7",
     "@types/node": "^22.7.8",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "@types/uuid": "^10.0.0",
     "@vitest/coverage-v8": "^2.1.3",
     "axios": "^1.7.7",
     "eslint": "^9.13.0",
+    "next": "^14.2.15",
+    "next-auth": "^5.0.0-beta.25",
     "prettier": "^3.3.3",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "typescript": "^5.5.3",
     "vite": "^5.4.8",
     "vite-plugin-dts": "^4.2.4",
     "vite-plugin-externalize-deps": "^0.8.0",
-    "vitest": "^2.1.3"
+    "vitest": "^2.1.3",
+    "zod": "^3.23.8"
   },
   "dependencies": {
     "jsonwebtoken": "^9.0.2",
     "uuid": "^10.0.0"
   },
   "peerDependencies": {
-    "axios": "*"
+    "axios": "*",
+    "next": "*",
+    "next-auth": "^5.0.0",
+    "react": "*",
+    "react-dom": "*",
+    "zod": "*"
   }
 }
